requirements.txt
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
boto3==1.39.9
botocore==1.39.9
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-environ==0.12.0
django-flatpickr==2.0.3
django-nested-admin==4.1.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
gunicorn==21.2.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
minio==7.2.15
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-monkey-business==1.1.0
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.13.1
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.31.0
uvloop==0.21.0
